+++ axum-extra/CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 # Unreleased
 
+- **added:** Add `JsonLines` for streaming newline delimited JSON ([#1093])
 - **change:** axum's MSRV is now 1.56 ([#1098])
 
+[#1093]: https://github.com/tokio-rs/axum/pull/1093
 [#1098]: https://github.com/tokio-rs/axum/pull/1098
 
 # 0.3.4 (08. June, 2022)